UseCase 작명에 대해 새로 합의한 내용을 기반으로 코드 수정 ({행동}{대상 리소스}UseCase)
